FEATURE: Implement ParsingTask for downloading and parsing the timeTable
